Fix format. Remove stuff from where it shouldn't be. Make testing syscalls optional. Remove a lot of `TODO`s. Use `Felt` instead of a marker type for the prime modulo meta.
