+++ src/executor.rs
@@ -484,44 +484,7 @@ impl<'a> ArgumentMapper<'a> {
             | (CoreTypeConcrete::Poseidon(_), _)
             | (CoreTypeConcrete::RangeCheck(_), _)
             | (CoreTypeConcrete::SegmentArena(_), _) => {}
-            (sierra_ty, arg_ty) => match sierra_ty {
-                CoreTypeConcrete::Array(_) => todo!("Array {arg_ty:?}"),
-                CoreTypeConcrete::Bitwise(_) => todo!("Bitwise {arg_ty:?}"),
-                CoreTypeConcrete::Box(_) => todo!("Box {arg_ty:?}"),
-                CoreTypeConcrete::EcOp(_) => todo!("EcOp {arg_ty:?}"),
-                CoreTypeConcrete::EcPoint(_) => todo!("EcPoint {arg_ty:?}"),
-                CoreTypeConcrete::EcState(_) => todo!("EcState {arg_ty:?}"),
-                CoreTypeConcrete::Felt252(_) => todo!("Felt252 {arg_ty:?}"),
-                CoreTypeConcrete::GasBuiltin(_) => todo!("GasBuiltin {arg_ty:?}"),
-                CoreTypeConcrete::BuiltinCosts(_) => todo!("BuiltinCosts {arg_ty:?}"),
-                CoreTypeConcrete::Uint8(_) => todo!("Uint8 {arg_ty:?}"),
-                CoreTypeConcrete::Uint16(_) => todo!("Uint16 {arg_ty:?}"),
-                CoreTypeConcrete::Uint32(_) => todo!("Uint32 {arg_ty:?}"),
-                CoreTypeConcrete::Uint64(_) => todo!("Uint64 {arg_ty:?}"),
-                CoreTypeConcrete::Uint128(_) => todo!("Uint128 {arg_ty:?}"),
-                CoreTypeConcrete::Uint128MulGuarantee(_) => todo!("Uint128MulGuarantee {arg_ty:?}"),
-                CoreTypeConcrete::Sint8(_) => todo!("Sint8 {arg_ty:?}"),
-                CoreTypeConcrete::Sint16(_) => todo!("Sint16 {arg_ty:?}"),
-                CoreTypeConcrete::Sint32(_) => todo!("Sint32 {arg_ty:?}"),
-                CoreTypeConcrete::Sint64(_) => todo!("Sint64 {arg_ty:?}"),
-                CoreTypeConcrete::Sint128(_) => todo!("Sint128 {arg_ty:?}"),
-                CoreTypeConcrete::NonZero(_) => todo!("NonZero {arg_ty:?}"),
-                CoreTypeConcrete::Nullable(_) => todo!("Nullable {arg_ty:?}"),
-                CoreTypeConcrete::RangeCheck(_) => todo!("RangeCheck {arg_ty:?}"),
-                CoreTypeConcrete::Uninitialized(_) => todo!("Uninitialized {arg_ty:?}"),
-                CoreTypeConcrete::Enum(_) => todo!("Enum {arg_ty:?}"),
-                CoreTypeConcrete::Struct(_) => todo!("Struct {arg_ty:?}"),
-                CoreTypeConcrete::Felt252Dict(_) => todo!("Felt252Dict {arg_ty:?}"),
-                CoreTypeConcrete::Felt252DictEntry(_) => todo!("Felt252DictEntry {arg_ty:?}"),
-                CoreTypeConcrete::SquashedFelt252Dict(_) => todo!("SquashedFelt252Dict {arg_ty:?}"),
-                CoreTypeConcrete::Pedersen(_) => todo!("Pedersen {arg_ty:?}"),
-                CoreTypeConcrete::Poseidon(_) => todo!("Poseidon {arg_ty:?}"),
-                CoreTypeConcrete::Span(_) => todo!("Span {arg_ty:?}"),
-                CoreTypeConcrete::StarkNet(_) => todo!("StarkNet {arg_ty:?}"),
-                CoreTypeConcrete::SegmentArena(_) => todo!("SegmentArena {arg_ty:?}"),
-                CoreTypeConcrete::Snapshot(_) => todo!("Snapshot {arg_ty:?}"),
-                CoreTypeConcrete::Bytes31(_) => todo!("Bytes31 {arg_ty:?}"),
-            },
+            (_, _) => todo!(),
         }
 
         Ok(())
