+++ src/libfuncs/bitwise.rs
@@ -1,6 +1,4 @@
 //! # Bitwise libfuncs
-//!
-//! TODO
 
 use super::{LibfuncBuilder, LibfuncHelper};
 use crate::{
