+++ src/libfuncs/gas.rs
@@ -1,6 +1,4 @@
 //! # Gas management libfuncs
-//!
-//! TODO
 
 use super::{LibfuncBuilder, LibfuncHelper};
 use crate::{
