+++ src/libfuncs/snapshot_take.rs
@@ -1,6 +1,4 @@
 //! # Snapshot taking libfuncs
-//!
-//! TODO
 
 use super::{LibfuncBuilder, LibfuncHelper};
 use crate::{
