+++ src/libfuncs/stark_net.rs
@@ -1,6 +1,4 @@
 //! # StarkNet libfuncs
-//!
-//! TODO
 
 use super::{LibfuncBuilder, LibfuncHelper};
 use crate::{
