+++ src/libfuncs/uint256.rs
@@ -1,6 +1,4 @@
 //! # `u256`-related libfuncs
-//!
-//! TODO
 
 use super::{LibfuncBuilder, LibfuncHelper};
 use crate::{
