+++ src/libfuncs/uint512.rs
@@ -1,6 +1,4 @@
 //! # `u512`-related libfuncs
-//!
-//! TODO
 
 use super::{LibfuncBuilder, LibfuncHelper};
 use crate::{
