+++ src/types/squashed_felt252_dict.rs
@@ -1,6 +1,4 @@
 //! # Squashed `Felt` dictionary type
-//!
-//! TODO
 
 use super::{TypeBuilder, WithSelf};
 use crate::{
